requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiogram==2.19
aiohttp==3.7.4
aiormq==3.3.1
aiosignal==1.2.0
APScheduler==3.7.0
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bidict==0.22.0
blis==0.7.9
boto3==1.26.1
botocore==1.29.1
CacheControl==0.12.11
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
confection==0.0.3
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
dask==2022.2.0
debugpy==1.6.3
decorator==4.4.2
Deprecated==1.2.13
deprecation==2.1.0
diskcache==5.2.1
dkpro-cassis==0.7.3
dm-tree==0.1.7
dnspython==1.16.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
fbmessenger==6.0.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==22.10.26
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gast==0.3.3
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.21.0
httptools==0.5.0
httpx==0.15.4
huggingface-hub==0.10.1
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.4.0
-e git+https://github.com/master1c/inception-external-recommender.git@3cbcd69a0cc252f4c128c6353df9536eb5eb8bac#egg=inception_external_recommender
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kafka-python==2.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
lightgbm==3.1.1
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mattermostwrapper==2.2
more-itertools==8.12.0
msgpack==1.0.4
multidict==5.2.0
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==2.5.1
nltk==3.7
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
parso==0.8.3
partd==1.3.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==2.0.10
protobuf==3.19.6
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
questionary==1.9.0
randomname==0.1.5
rasa==2.8.0
rasa-sdk==2.8.6
redis==3.5.3
regex==2021.7.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
rust-fst==0.1.2
s3transfer==0.6.0
sacremoses==0.0.53
sanic==20.12.7
Sanic-Cors==0.10.0.post3
sanic-jwt==1.8.0
Sanic-Plugins-Framework==0.9.5
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentence-transformers==0.3.9
sentencepiece==0.1.91
sentry-sdk==1.2.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.42
srsly==2.4.5
stack-data==0.6.0
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.13.0
tensorflow-text==2.3.0
termcolor==2.1.0
terminaltables==3.1.10
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.1
toolz==0.12.0
toposort==1.7
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
twilio==6.50.1
typeguard==2.13.3
typer==0.4.2
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzdata==2022.6
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.12
uvloop==0.14.0
wasabi==0.10.1
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
